separate out table from main view
