+++ packages/common/src/elements/CloseButton.svelte
